Последний рывок

+++ AvitoMonitor/Form1.cs
@@ -16,41 +16,40 @@ using CsQuery;
 
 namespace AvitoMonitor{
     public partial class AvitoMonitor : Form{
-        string MainLink = @"https://www.avito.ru";
         string WAY = @"AvitoMonitor.db";
-        string PATHtoIMG = @"Image Maker";
-        string conectString = "Data Source=AvitoMonitor.db;Version=3;New=False;Compress=True;";
-        string SEARCHstring = "https://www.avito.ru";
-        string htmlCode;
+        string WAY_FOR_NEW_ADDS = @"AvitoMonitorNew.db";
+        string WAY_FOR_SUPPORT_DB = @"AvitoMonitorSupport.db";
         public AvitoMonitor(){
             InitializeComponent();
             Support.CreationOfDataBase(WAY);
-            Support.CreationOfDirectory(PATHtoIMG);
+            Support.CreationOfDataBaseForNewAdds(WAY_FOR_NEW_ADDS);
+            Support.CreationOfDirectory(Support.PATHtoIMG);
         }
 
         private void button1_Click(object sender, EventArgs e){
             try { 
                 using (WebClient client = new WebClient()){
+                    //Проверка комбобоксов
                     if (comboBoxForCity.Text.Trim() != "") { 
-                        SEARCHstring += "/" + Support.Cities[comboBoxForCity.Text];    
+                        Support.SEARCHstring += "/" + Support.Cities[comboBoxForCity.Text];    
                     } else {
                         MessageBox.Show("Ошибка поиска, город не выбран",
                         "Вы должны выбрать город", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                        SEARCHstring = "https://www.avito.ru";
+                        Support.SEARCHstring = "https://www.avito.ru";
                         return;
                     }
-                    if (comboBoxType.Text.Trim() != "") { 
-                        SEARCHstring += "/" + Support.Types[comboBoxType.Text];
+                    if (comboBoxType.Text.Trim() != "") {
+                        Support.SEARCHstring += "/" + Support.Types[comboBoxType.Text];
                     } else {
                         MessageBox.Show("Ошибка поиска, категория не выбрана",
                         "Вы должны выбрать тип(категорию)", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                        SEARCHstring = "https://www.avito.ru";
+                        Support.SEARCHstring = "https://www.avito.ru";
                         return;
                     }
-                    if (searchtextBox.Text.Replace(" ", "") != "") { 
-                        SEARCHstring += "?s_trg=11&q=" + searchtextBox.Text;
+                    if (searchtextBox.Text.Replace(" ", "") != "") {
+                        Support.SEARCHstring += "?s_trg=11&q=" + searchtextBox.Text;
                     } else if (searchtextBox.Text.Replace(" ", "") == "") {
-                        SEARCHstring += "?s_trg=11";
+                        Support.SEARCHstring += "?s_trg=11";
                     } /*else {
                         MessageBox.Show("Ошибка поиска, текст поиска отсутствует",
                         "Вы должны ввести текст поиска", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
@@ -58,27 +57,34 @@ namespace AvitoMonitor{
                         return;
                     }*/
                     
+                    //создание БД и деректории для картинок при их отсутствии
                     Support.CreationOfDataBase(WAY);
-                    Support.CreationOfDirectory(PATHtoIMG);
+                    Support.CreationOfDataBaseForNewAdds(WAY_FOR_NEW_ADDS);
+                    Support.CreationOfDirectory(Support.PATHtoIMG);
                     richTextBox1.AppendText("Downloading... ");
-                    try { 
-                        htmlCode = client.DownloadString(SEARCHstring);
+                    //Проверка на существование страницы
+                    try {
+                        Support.htmlCode = client.DownloadString(Support.SEARCHstring);
                     } catch {
-                        richTextBox1.AppendText(SEARCHstring);
+                        richTextBox1.AppendText(Support.SEARCHstring);
                         return;
                     }
                 
                     richTextBox1.AppendText("ok.\n");
 
+                    string sourceFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, WAY_FOR_NEW_ADDS);
+                    string destFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, WAY_FOR_SUPPORT_DB);
+                    File.Copy(sourceFile, destFile, true);
+
                     Encoding utf8 = Encoding.GetEncoding("UTF-8");
                     Encoding win1251 = Encoding.GetEncoding("Windows-1251");
 
-                    byte[] utf8Bytes = win1251.GetBytes(htmlCode);
+                    byte[] utf8Bytes = win1251.GetBytes(Support.htmlCode);
                     byte[] win1251Bytes = Encoding.Convert(utf8, win1251, utf8Bytes);
 
-                    htmlCode = win1251.GetString(win1251Bytes);
+                    Support.htmlCode = win1251.GetString(win1251Bytes);
 
-                    CQ dom = htmlCode;
+                    CQ dom = Support.htmlCode;
                     CQ items = dom.Select(".item");
                     int j = 0;
                     CQ title = dom.Select(".item-description-title-link");
@@ -89,7 +95,7 @@ namespace AvitoMonitor{
                         if (item == null || title == null || itemphoto == null || itemtime == null) {
                             MessageBox.Show("Ошибка поиска, По вашему запросу ничего не найдено",
                             "Ничего не найдено", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                            SEARCHstring = "https://www.avito.ru";
+                            Support.SEARCHstring = "https://www.avito.ru";
                             return;
                         }
 
@@ -111,7 +117,7 @@ namespace AvitoMonitor{
                                 item[".price"].Text().Replace("?", "").Replace(" ", "") != "" &&
                                 itemtime[i]["data-relative-date"] != "" && title[i]["href"] != "") {
                                 string path = AppDomain.CurrentDomain.BaseDirectory +
-                                        @"\" + PATHtoIMG + @"\" + "img.jpg";
+                                        @"\" + Support.PATHtoIMG + @"\" + "img.jpg";
                                 using (WebClient clientimg = new WebClient()){
 
                                     client.DownloadFile(Support.ImageString + itemphoto[j]["src"], 
@@ -126,7 +132,6 @@ namespace AvitoMonitor{
                                     string imgName = Path.GetFileName(path).Replace(Path.GetExtension(path), ""); // запишем в переменную имя файла, не забыв удалить расширение с точкой, получим «img»
 
                                     // записываем информацию в базу данных
-
 
                                     using (SQLiteConnection Connect = new SQLiteConnection(
                                         @"Data Source=" + AppDomain.CurrentDomain.BaseDirectory + @"\" 
@@ -143,12 +148,39 @@ namespace AvitoMonitor{
                                         Command.Parameters.AddWithValue("@type", comboBoxType.Text);
                                         Command.Parameters.AddWithValue("@image", _imageBytes); // присваиваем переменной значение
                                         Command.Parameters.AddWithValue("@format", imgFormat);
-                                        Command.Parameters.AddWithValue("@add_link", MainLink + title[i]["href"]);
+                                        Command.Parameters.AddWithValue("@add_link", Support.MainLink + title[i]["href"]);
                                         Connect.Open();
                                         Command.ExecuteNonQuery();
                                         Connect.Close();
+                                    }
+                                    using (SQLiteConnection connection = new SQLiteConnection(@"Data Source=" + 
+                                        AppDomain.CurrentDomain.BaseDirectory + @"\" + WAY_FOR_NEW_ADDS + @"; Version=3;")) {
+                                        string CommandText = "SELECT count(Цена) " +
+                                            "FROM AvitoMonitorNew WHERE name='" + title[i]["title"] + @"'" +
+                                            @"AND Цена='" + item[".price"].Text().Replace("?", "P") + @"'" +
+                                            @"AND Опубликовано='" + itemtime[i]["data-relative-date"] + @"'" +
+                                            @"AND Картинка='" + _imageBytes + @"'"+
+                                            @"AND Ссылка на Объявление='" + Support.MainLink + title[i]["href"] + @"'"; // Если этот запрос не выполняется
+                                        SQLiteCommand Command = new SQLiteCommand(CommandText, connection);
+                                        int countRows = (int)Command.ExecuteScalar();
 
-
+                                        if (countRows == 0){
+                                            CommandText = "INSERT INTO [AvitoMonitorNew] ([Время], [Название], [Цена], [Город], [Опубликовано], [Тип], [Картинка], [Формат картинки], [Ссылка на Объявление]) " +
+                                            "VALUES(@time, @name, @price, @city, @addtime, @type, @image, @format, @add_link)";
+                                            Command = new SQLiteCommand(CommandText, connection);
+                                            Command.Parameters.AddWithValue("@time", DateTime.Now.ToString());
+                                            Command.Parameters.AddWithValue("@name", title[i]["title"]);
+                                            Command.Parameters.AddWithValue("@price", item[".price"].Text().Replace("?", "P"));
+                                            Command.Parameters.AddWithValue("@city", comboBoxForCity.Text);
+                                            Command.Parameters.AddWithValue("@addtime", itemtime[i]["data-relative-date"]);
+                                            Command.Parameters.AddWithValue("@type", comboBoxType.Text);
+                                            Command.Parameters.AddWithValue("@image", _imageBytes); 
+                                            Command.Parameters.AddWithValue("@format", imgFormat);
+                                            Command.Parameters.AddWithValue("@add_link", Support.MainLink + title[i]["href"]);
+                                            connection.Open();
+                                            Command.ExecuteNonQuery();
+                                            connection.Close();
+                                        }
                                     }
                                     _fileStream.Close();
                                     File.Delete(path);
@@ -161,11 +193,12 @@ namespace AvitoMonitor{
 
                             Support.ImageStringDefault();
                         } catch (Exception ex){ 
-                            //MessageBox.Show(ex.Message);
+                            MessageBox.Show(ex.Message);
                             continue;
                         }
                         richTextBox1.AppendText("\n");
                     }
+                    File.Delete(WAY_FOR_SUPPORT_DB);
                 }
             } catch (Exception) {
                 MessageBox.Show("Нет соеденения с интернетом",
@@ -173,7 +206,7 @@ namespace AvitoMonitor{
                 return;
             }
         }
-        void ExecuteQuery(string txtQuery) {
+        /*void ExecuteQuery(string txtQuery) {
             using(SQLiteConnection connection = new SQLiteConnection(conectString)) {
                 SQLiteCommand command = new SQLiteCommand();
                 command = connection.CreateCommand();
@@ -184,9 +217,9 @@ namespace AvitoMonitor{
                 command.Dispose();
                 command = null;
             }
-        }
+        }*/
         void LoadData() {
-            using(SQLiteConnection connection = new SQLiteConnection(conectString)) {
+            using(SQLiteConnection connection = new SQLiteConnection(Support.conectString)) {
                 SQLiteCommand command = connection.CreateCommand();
                 string ComandText = "select *from AvitoMonitor";
                 SQLiteDataAdapter DB = new SQLiteDataAdapter(ComandText, connection);
@@ -196,10 +229,21 @@ namespace AvitoMonitor{
                 DataTable DT;
                 DT = DS.Tables[0];
                 dataGridView1.DataSource = DT;
-                DT = null;
-                DS = null;
-                DB = null;
-                command = null;
+                connection.Close();
+                connection.Dispose();
+            }
+        }
+        void LoadNewData() {
+            using(SQLiteConnection connection = new SQLiteConnection(Support.conectStringForNewDB)) {
+                SQLiteCommand command = connection.CreateCommand();
+                string ComandText = "select *from AvitoMonitorNew";
+                SQLiteDataAdapter DB = new SQLiteDataAdapter(ComandText, connection);
+                DataSet DS = new DataSet();
+                DS.Reset();
+                DB.Fill(DS);
+                DataTable DT;
+                DT = DS.Tables[0];
+                dataGridView1.DataSource = DT;
                 connection.Close();
                 connection.Dispose();
             }
@@ -212,9 +256,33 @@ namespace AvitoMonitor{
         }
 
         private void loadbd_Click(object sender, EventArgs e){
+            dataGridView1.DataSource = null;
+            GC.Collect();
+            GC.WaitForPendingFinalizers();
+
             if (!File.Exists(WAY))
                 Support.CreationOfDataBase(WAY);
             LoadData();
+
+            dataGridView1.AutoSizeColumnsMode =
+                DataGridViewAutoSizeColumnsMode.AllCells;
+            dataGridView1.AutoResizeColumns();
+
+            dataGridView1.AutoSizeRowsMode =
+                DataGridViewAutoSizeRowsMode.AllCells;
+            dataGridView1.AutoResizeRows(
+                DataGridViewAutoSizeRowsMode.AllCellsExceptHeaders);
+        }
+
+        private void loadnewdb_Click_1(object sender, EventArgs e){
+            dataGridView1.DataSource = null;
+            GC.Collect();
+            GC.WaitForPendingFinalizers();
+
+            if (!File.Exists(WAY_FOR_NEW_ADDS))
+                Support.CreationOfDataBaseForNewAdds(WAY_FOR_NEW_ADDS);
+            LoadNewData();
+
             dataGridView1.AutoSizeColumnsMode =
                 DataGridViewAutoSizeColumnsMode.AllCells;
             dataGridView1.AutoResizeColumns();
@@ -258,12 +326,13 @@ namespace AvitoMonitor{
         }
 
         private void DeleteDB_Click(object sender, EventArgs e){
-            if (File.Exists(WAY)){
+            if (File.Exists(WAY) && File.Exists(WAY_FOR_NEW_ADDS)){
                 try { 
                     dataGridView1.DataSource = null;
                     GC.Collect();
                     GC.WaitForPendingFinalizers();
                     Support.DeleteDB(WAY);
+                    Support.DeleteDB(WAY_FOR_NEW_ADDS);
                     MessageBox.Show("База данных успешно удалена", "Удаление прошло успешно",
                         MessageBoxButtons.OK, MessageBoxIcon.None);
                 } catch (Exception ex) { 
@@ -277,26 +346,47 @@ namespace AvitoMonitor{
         }
 
         private void копироватьБазуДанныхToolStripMenuItem_Click(object sender, EventArgs e){
-            if (true){
-                SaveFileDialog savedialog = new SaveFileDialog();
-                savedialog.InitialDirectory = @"C:\";
-                savedialog.Title = "Сохранить Базу Данных как...";
-                savedialog.OverwritePrompt = true;
-                savedialog.CheckPathExists = true;
-                savedialog.Filter = "DataBase files(*.db)|*.db";
-                savedialog.ShowHelp = true;
-                if (savedialog.ShowDialog() == DialogResult.OK) { 
-                    try{
-                        string sourcePath = AppDomain.CurrentDomain.BaseDirectory;
-                        string targetPath = Path.GetFullPath(savedialog.FileName);
-                        string sourceFile = Path.Combine(sourcePath, WAY);
-                        string destFile = Path.Combine(targetPath, savedialog.FileName);
-                        File.Copy(sourceFile, destFile, true);
-                    } catch (Exception ex){
-                        MessageBox.Show("Невозможно сохранить базу данных, возможно вы её удалили \n" 
-                            + ex.Message, "Ошибка",
-                        MessageBoxButtons.OK, MessageBoxIcon.Error);
-                    }
+            SaveFileDialog savedialog = new SaveFileDialog();
+            savedialog.InitialDirectory = @"C:\";
+            savedialog.Title = "Сохранить Базу Данных как...";
+            savedialog.OverwritePrompt = true;
+            savedialog.CheckPathExists = true;
+            savedialog.Filter = "DataBase files(*.db)|*.db";
+            savedialog.ShowHelp = true;
+            if (savedialog.ShowDialog() == DialogResult.OK) { 
+                try{
+                    string sourcePath = AppDomain.CurrentDomain.BaseDirectory;
+                    string targetPath = Path.GetFullPath(savedialog.FileName);
+                    string sourceFile = Path.Combine(sourcePath, WAY);
+                    string destFile = Path.Combine(targetPath, savedialog.FileName);
+                    File.Copy(sourceFile, destFile, true);
+                } catch (Exception ex){
+                    MessageBox.Show("Невозможно сохранить базу данных, возможно вы её удалили \n" 
+                        + ex.Message, "Ошибка",
+                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+            }
+        }
+
+        private void копироватьБазуДанныхДляНовыхОбъявленийToolStripMenuItem_Click(object sender, EventArgs e){
+            SaveFileDialog savedialog = new SaveFileDialog();
+            savedialog.InitialDirectory = @"C:\";
+            savedialog.Title = "Сохранить Базу Данных как...";
+            savedialog.OverwritePrompt = true;
+            savedialog.CheckPathExists = true;
+            savedialog.Filter = "DataBase files(*.db)|*.db";
+            savedialog.ShowHelp = true;
+            if (savedialog.ShowDialog() == DialogResult.OK) { 
+                try{
+                    string sourcePath = AppDomain.CurrentDomain.BaseDirectory;
+                    string targetPath = Path.GetFullPath(savedialog.FileName);
+                    string sourceFile = Path.Combine(sourcePath, WAY_FOR_NEW_ADDS);
+                    string destFile = Path.Combine(targetPath, savedialog.FileName);
+                    File.Copy(sourceFile, destFile, true);
+                } catch (Exception ex){
+                    MessageBox.Show("Невозможно сохранить базу данных, возможно вы её удалили \n" 
+                        + ex.Message, "Ошибка",
+                    MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
             }
         }
